no adbyby no update.

+++ luci-app-adbyby-plus/root/usr/share/adbyby/adblock.sh
@@ -1,5 +1,8 @@
 #!/bin/sh
 
+enable=$(uci -q get adbyby.@adbyby[0].enable)
+[ "$enable" == "1" ] || exit 0
+
 rm -f /tmp/adbyby.updated /tmp/adbyby/admd5.json
 
 sh /usr/share/adbyby/rule-update
